dating_site/app/requirements.txt
Django==3.2.12
gunicorn==20.1.0
psycopg2-binary==2.9.1
Pillow==9.0.0
djangorestframework==3.13.1
django-filter==21.1
geopy==2.2.0
celery==5.2.2
redis==3.5.3
flower==1.0.0
drf-yasg==1.20.0
requests==2.26.0
PyJWT==2.1.0
